requirements.txt
tensorflow==2.16.1
Pillow==10.2.0
numpy==1.26.4
pandas==2.2.1
matplotlib==3.8.3
scikit-learn==1.4.1
seaborn==0.13.2
